Split message components in multiple files
